renamed bindings to qpSWIFT_sparse_bindings
